Add new exception raised on API key being rejected by the backend.
We now know wehn a token is explicitly rejected as invalid, as opposed
to any other exchange error.

This commit also removes dependency on the `backoff` package and
its use when requesting token. The backoff mechanism was used only
in a single place, and all the client libraries use their own
backoff anyway.